+++ components/event-publisher-proxy/pkg/eventmesh/types.go
@@ -1,4 +1,4 @@
-package ems
+package eventmesh
 
 type Qos string
